APP_epa_gauges_fetcher/requirements.txt
requests~=2.26.0
python-dateutil~=2.8.2
git+https://github.com/dbw16/riverLevelsModual.git
